requirements.txt
asgiref==3.9.1
click==8.3.0
dj-database-url==3.0.1
Django==5.2.6
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.11
python-decouple==3.8
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.15.0
uvicorn==0.37.0
whitenoise==6.11.0
